:sparkles: 설문 응답 저장 로직 개선 및 만족도 점수 처리 방식 수정

+++ app/src/main/java/net/focustation/myapplication/ui/screen/survey/SurveyScreen.kt
@@ -443,7 +443,7 @@ private fun rememberSurveyPages(viewModel: SurveyViewModel): List<SurveyPage> =
                             ),
                         selectedValue = state.satisfactionScore.toString(),
                         onSelected = {
-                            viewModel.setSatisfactionScore(it.toFloat())
+                            viewModel.setSatisfactionScore(it.toInt())
                             goNext()
                         },
                     )
